(PC-43513)[API] chore: updates from review
- split tests in test_api.py
- check presence in DB for tests
